[rfc] Generate the Method enum and parse_message function.
A stab at #45.
This uses a recursive macro to build up both things simultaneously.
That does mean that with more messages, we can hit the macro expansion
limit, but I don't think you can do the filtering without recursion.

This doesn't do anything for LsService::handle_message (yet) but it
should be doable.

I manually checked that the `--pretty expanded` output is equivalent.